requirements.txt
Flask==2.1.0
flask_sqlalchemy
mysql-connector-python
werkzeug==2.0.2
sqlalchemy==1.4.49
python-dotenv
pymysql
gunicorn
twilio==9.7.2
cryptography
Flask-JWT-Extended
PyJWT 
pytest
pytest-mock

